requirements.txt
# Worldbuilding Interactive Program -- Dependencies
# Core engine
pydantic>=2.0
networkx>=3.0

# Desktop application (Phase 3)
PySide6-Essentials>=6.7.0
qt-material>=2.14
qtawesome>=1.3.0
qasync>=0.27.0
platformdirs>=4.0

# Development
pytest>=7.0
